refactored how attendance is read from tournaments
